1. Edit Transaction Dialog - Added option of "Duplicate" in the "Status" drop down list.  When this wasn't present, after saving a transaction with a previous status of "Duplicate", it would change to "None" on the save, even if not editing "Status".
2. Universal Import - When removing a field, no item was selected
after the remove.  Now when removing a field, the next item after the
deleted field is selected so that fields can be quickly removed
without having to reselect after hitting the Remove button.

3. Deleting Transaction in Bank Account - There was a bug when
deleting a transaction, the screen would not refresh correctly with
the right underlying data, so the transactions were in the wrong
locations.  Now a refresh of the underlying data occurs and only
transactions after the deleted transactions are refreshed.  Also, the
top visible transaction index is stored and then set on the
"EnsureVisible" method so the list doesn't put the deleted transaction
selected index at the top, but instead tries to keep the list in the
same location before the Delete.  I did not change this for New or
Edit transaction since the inserting or editing the transaction might
change the location of where it falls in the list based on the
currently selected sorting type.

git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@224 7f5d68d0-c712-0410-839f-9b116a576cc5